llava/requirements.txt
# Core dependencies for GCAV implementation
torch>=2.0.0
transformers>=4.35.0
Pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0
requests>=2.28.0
pyyaml>=6.0
tqdm>=4.64.0
matplotlib>=3.5.0
pandas>=1.5.0

# Optional for advanced evaluations  
nltk>=3.8
rouge-score>=0.1.2
